Kernel: Remove problematic memory mapping methods
mmap() & mmap_region() methods are removed from ACPI & DMI components,
and we replace them with the new MM.allocate_kernel_region() helper.

Instead of doing a raw calculation for each VM address, from now on we
can use helper functions to do perform those calculations in a neat,
reusable and readable way.

+++ Kernel/PCI/Access.h
@@ -14,7 +14,7 @@ public:
     virtual u32 get_BAR4(Address address) { return read32_field(address, PCI_BAR4); }
     virtual u32 get_BAR5(Address address) { return read32_field(address, PCI_BAR5); }
 
-    virtual u32 get_BAR_Space_Size(Address address, u8 bar_number)
+    virtual size_t get_BAR_Space_Size(Address address, u8 bar_number)
     {
         // See PCI Spec 2.3, Page 222
         ASSERT(bar_number < 6);
